tests/wpt/web-platform-tests/html/semantics/forms/the-option-element/option-text-setter.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form>
    <select>
        <option value='foo'>bar</option>
    </select>
</form>
<script>
test(() => {
    var option = document.querySelector('option');
    var textChild = option.firstChild;
    assert_equals(textChild.nodeValue, "bar", "Verify that text child node's value equals the option value.");
    assert_true(textChild.isConnected, 'Verify that text child node is in the document.');
    option.text = "baz";
    assert_equals(textChild.nodeValue, "bar", 'Verify that the text child node does not have an updated value.');
    assert_false(textChild.isConnected, 'Verify that the text child node is not in the document.');
    assert_false(textChild == option.firstChild, 'Verify that text child node was replaced by a different text child node.');
    assert_equals(option.firstChild.nodeValue, "baz", 'Verify that the new text child node does equal the updated value.');
    assert_equals(option.text, "baz", 'Verify that option text getter returns the updated value.');
    option.text = "";
    assert_true(option.firstChild == null, 'Verify that after setting to empty string there are no child text nodes.');
}, 'Verify that using HTMLOptionElement.text setter does not update the existing text child node.');
</script>
