Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-option-element/option-text-setter-expected.txt



FAIL Verify that using HTMLOptionElement.text setter does not update the existing text child node. assert_equals: Verify that the text child node does not have an updated value. expected "bar" but got "baz"
